requirements.txt
emoji==2.14.0
nltk==3.2.2
pandas==2.2.3
scikit-learn==1.6.0
pyspellchecker==0.8.2
parsivar==0.2.2
spacy==3.8.3